camera_keypoints_monitor/pyproject.toml
[tool.poetry]
name = "camera-keypoints-monitor"
version = "0.1.0"
description = "CameraKeypoints"
authors = ["Kamyar Ghiam <kamyarghiam@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
supabase = "^2.20.0"
slack-sdk = "^3.36.0"
opencv-python-headless = "^4.12.0.88"
boto3 = "^1.40.48"
pillow = "^11.3.0"
requests = "^2.32.5"
pyyaml = "^6.0.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"